If process already running just reconnect any auto-connect configs
When a second instance is started it requests the running instance
to connect any configs set as auoto-connect --- if such configs are
already connected the status window(s) of those connections pops up.

If no configs are set to auto-connect, the behaviour is unchanged
(i.e, the "already running" popup appears).

Signed-off-by: Selva Nair <selva.nair@gmail.com>